requirements.txt
Django==3.2
djangorestframework==3.14.0
drf-yasg==1.21.5
pandas==2.0.3
Pillow==10.0.0
pytrends==4.9.2
requests==2.31.0
urllib3==1.26.18